feat: adds support for robbers gamemode
This commit adds support for the robbers gamemode. In the robbers
gamemode, you will be allowed to specify a number of robbers,
defaulting to 1, and specify various robber-specific hints for chests.
Various content-specific hints were abstracted into a more generic
form, and some tests were added to make the process of adding new hints
less likely to run into problems. A new process was also added to
allow modifying the state when switching game modes, to allow removing
content or hints that are invalid for the new gamemode when switching.