feat: adds support for "confuse" gamemode
This change adds support for the final gamemode in Mimic Logic,
Confuse. The number of confused chests is hard-coded as one,
as this matches the ruleset in-game, and for now "confused chests"
are not shown anywhere in the UI nor reported on in solutions.
Solution tile has also been updated to always highlight solutions
that contain a mimic, and to sort the mimic solution highest.
This helps prevent accidentally opening a mimic chest.

fix: fixed hints for rankMoreMimics and rankSameMimics