feat: reset button now resets optional contents
When solving a puzzle it may make sense to provide exact values for
the various non-mimic content types. However, these values change
between puzzles almost every time. To provide a clean slate for a
new puzzle, it makes sense to allow the reset button to reset
these values as well.