feat: change chest color select to radio
This update changes the select dropdown for chest color to a radio instead
to help reduce the number of clicks needed when setting up a puzzle.
The "radio tiles" were also abstracted and re-used in the mini-chest select,
in the chest layout selection, and in the new color selector. Some minor
changes were also made to chest tiles to reduce layout shift by providing
a minimum height for the hint (approximately 3 lines high).