feat: improve direction / rank selection
This change re-orders direction and rank selectors for hints to be
specifically ordered rather than defined by the order of the underlying
enum. This chnage also adds logic to filter these selections to only
show valid direction and rank options for a given chest and grid size.
Note: It is still possible to have an invalid direction or rank selected
by selecting an option and then changing the grid size such that the
option becomes invalid. However, it will be easy to change and not
appear glitchy. (Selected option will always be in list, until changed)